requirements.txt
Flask~=3.0.0
google-auth-oauthlib~=1.2.0
google-cloud-firestore~=2.16.0
google-cloud-storage~=2.16.0
google-api-python-client~=2.130.0
